correctting erros on PIO build

+++ Linefollower/include/config.h
@@ -71,6 +71,6 @@ static const uint8_t UNGRIP_POS = 180;
 static const uint8_t LIFT_POS = 0;
 static const uint8_t LOWER_POS = 180;
 
-static const uint8_t GRIP_DELAY = 500; // ms to wait after gripping/ungripping
-static const uint8_t LOWER_DELAY = 500; // ms to wait after lowering
-static const uint8_t RAISE_DELAY = 500; // ms to wait after raising
+static const uint16_t GRIP_DELAY = 500; // ms to wait after gripping/ungripping
+static const uint16_t LOWER_DELAY = 500; // ms to wait after lowering
+static const uint16_t RAISE_DELAY = 500; // ms to wait after raising
